+++ test/utils/getUserOpHash.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.13;
 import {UserOperation} from "src/external/UserOperation.sol";
 
 /// @notice Get the userOperation hash over a user operation, entryPoint and chainId
-function getUserOpHash(UserOperation memory userOp, address entryPoint, uint8 chainId) pure returns (bytes32) {
+function getUserOpHash(UserOperation memory userOp, address entryPoint, uint256 chainId) pure returns (bytes32) {
     bytes32 userOpHash = keccak256(
         abi.encode(
             userOp.sender,
